Move helper functions to standalone file
